Renamed helper files/dirs so they're more explicit in their use.
