+++ src/javascripts/lib/handlebars-helpers/is.js
