+++ src/javascripts/status-bar.js
@@ -18,7 +18,7 @@
 	// Use handlebars for templating.
 	var handlebars = require('handlebars');
 	// Register some helpers for smarter templates.
-	handlebars.registerHelper(require('./lib/helpers'));
+	handlebars.registerHelper(require('./lib/handlebars-helpers'));
 
 	/*
 	 *
